feat(manager-react-components): migrate hook date folder
ref: #MANAGER-19189
