+++ packages/manager-react-components/src/hooks/date/useDateFnsLocale/index.ts
@@ -1,0 +1,2 @@
+export { useDateFnsLocale } from './useDateFnsLocale';
+export type { LocaleKey, LOCALE_MAP } from './useDateFnsLocale.type';
