+++ packages/manager-react-components/src/hooks/date/useFormatDate/index.ts
@@ -1,0 +1,5 @@
+export { useFormatDate } from './useFormatDate';
+export type {
+  FormatDateOptions,
+  DEFAULT_UNKNOWN_DATE_LABEL,
+} from './useFormatDate.type';
