+++ packages/manager-react-components/src/hooks/date/useFormatDate/useFormatDate.type.ts
@@ -1,9 +1,3 @@
-import { useCallback } from 'react';
-import { format as formatDateFns, isValid } from 'date-fns';
-import { useDateFnsLocale } from './useDateFnsLocale';
-
-export const DEFAULT_UNKNOWN_DATE_LABEL = 'N/A';
-
 /**
  * Format options for date strings
  *
@@ -28,27 +22,4 @@ export type FormatDateOptions = {
   invalidDateDisplayLabel?: string;
 };
 
-export const useFormatDate = ({
-  invalidDateDisplayLabel = DEFAULT_UNKNOWN_DATE_LABEL,
-}: { invalidDateDisplayLabel?: string } = {}) => {
-  const locale = useDateFnsLocale();
-
-  const formatDate = useCallback(
-    ({ date, format = 'PP' }: FormatDateOptions): string => {
-      const parsedDate = typeof date === 'string' ? new Date(date) : date;
-
-      if (!parsedDate || !isValid(parsedDate)) {
-        return invalidDateDisplayLabel;
-      }
-
-      try {
-        return formatDateFns(parsedDate, format, { locale });
-      } catch (_e) {
-        return invalidDateDisplayLabel;
-      }
-    },
-    [locale, invalidDateDisplayLabel],
-  );
-
-  return formatDate;
-};
+export const DEFAULT_UNKNOWN_DATE_LABEL = 'N/A';
